Cash: Double to BigDecimal refactoring for amount

+++ cash/src/main/java/ru/yandex/practicum/cash/dto/CashOperationRequest.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
@@ -13,5 +15,6 @@ public class CashOperationRequest {
     private Long userId;
     private String operation; // "DEPOSIT" или "WITHDRAW"
     private String currency;
-    private Double amount;
+    private BigDecimal amount;
+
 }
